+++ R/task_steadystate.R
@@ -49,7 +49,7 @@ runSteadyState <- function(calculate_jacobian = NULL, perform_stability_analysis
     if (do_method)
       set_method_settings(method_settings, c_method)
     
-    grab_msg(c_model$compileIfNecessary())
+    compile_and_check(c_model)
     
     # initialize task
     assert_that(