+++ R/task_timecourse.R
@@ -65,7 +65,7 @@ runTimeCourse <- function(duration = NULL, dt = NULL, intervals = NULL, suppress
     if (do_method)
       set_method_settings(method_settings, c_method)
     
-    grab_msg(c_model$compileIfNecessary())
+    compile_and_check(c_model)
     
     # initialize task
     assert_that(