+++ NEWS.md
@@ -1,6 +1,10 @@
-# CoRC 0.0.9001
+# CoRC 0.0.1.9001
 
-* fix errors with argument 'type' of setCompartments() and setGlobalQuantities()
+* Add model event functioniality (`event()`, `event_strict()`, `getEvents()`, `setEvents()`, `deleteEvent()`)
+* Allow expressions to be automatically coerced from finite numeric and logical values.
+* Fix crashes on entity deletion. Deletion functions now delete recursively.
+* Generally warn if model fails to compile.
+* Fix errors with argument` type` of `setCompartments()` and `setGlobalQuantities()`
 
 # CoRC 0.0.1